Conflicts:
	systems/ActionMovementSystem.hh
	systems/LifeSystem.hh
	systems/MoveForwardSystem.hh
	systems/MoveSequenceSystem.hh
	systems/SFMLDisplaySystem.cpp

+++ components/TeamComponent.cpp
@@ -34,7 +34,7 @@ void	TeamComponent::deserializeFromFileSpecial(const std::string &lastline, std:
   (void)input;
 
   if (std::regex_match(lastline, std::regex("team=.+")))
-    this->_team = std::stof(lastline.substr(5));
+    this->_team = std::stoul(lastline.substr(5));
   else
     throw EntityFileException("Bad argument : \"" + lastline + "\"", lineno);
 }
